pyproject.toml
[project]
name = "job-scout"
version = "0.1.0"
description = "Lightweight job scraping and alerting"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27,<1.0",
    "beautifulsoup4>=4.12,<5.0",
    "pydantic>=2.6,<3.0",
    "typer>=0.12,<1.0",
    "rich>=13.0,<14.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
tls = [
    "curl_cffi>=0.7,<1.0",
]
bot = [
    "google-genai>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[project.scripts]
job-scout = "job_scout.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_scout"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
src = ["src"]
